Add Typescript and guesses
