+++ src/Components/App/App.tsx
@@ -1,0 +1,49 @@
+import '../App/App.css'
+import React, { useState, useEffect } from 'react'
+import getData from "../../apicalls"
+
+type data = {
+  data: string[]
+}
+
+interface Guesses {
+  Americas: number,
+  Asia: number,
+  Oceania: number,
+  Europe: number,
+  Africa: number
+}
+
+const App: React.FC = () => {
+  const [data, setData] = useState([])
+  const [correctGuesses, setCorrectGuesses] = useState<Guesses>({ Americas: 0, Asia: 8, Oceania: 0, Europe: 0, Africa: 0 })
+  const [incorrectGuesses, setIncorrectGuesses] = useState<Guesses>({ Americas: 0, Asia: 4, Oceania: 0, Europe: 0, Africa: 0 })
+  const initApp = async () => {
+    try {
+      const response = await getData()
+      setData(response.data.continents)
+    }
+    catch (error) {
+      console.log(error)
+    }
+  }
+  useEffect(() => {
+    initApp()
+    console.log("data", data)
+    keepScore('Asia')
+  }, [])
+
+  const keepScore = (continent: string) => {
+    const total = correctGuesses[continent] + incorrectGuesses[continent]
+    const score = (correctGuesses[continent]/total * 100).toFixed() + '%'
+    console.log('score', score)
+  }
+
+  return (
+    <main className="app-container">
+      <h1>Trivia Game</h1>
+    </main>
+  )
+}
+
+export default App
